requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.3.0
Django==4.0
django-browser-reload==1.11.0
-e git+https://github.com/jazzband/django-debug-toolbar.git@7ab6b0f4802f19a28e4141df0be70b4f0b28eae1#egg=django_debug_toolbar
django-filter==23.2
django-tailwind==3.6.0
django-tinymce==3.6.1
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pillow==10.0.0
psycopg2-binary==2.9.6
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
